Modularized gcode interpreter
